Move webmock monkey patch into single file
doesn't need to be all nested for no reason
